不要コード削除
