+++ pages/index.vue
@@ -88,10 +88,8 @@
           title="市町村別感染状況"
           :title-id="'patients-per-cities'"
           :date="Data.patients.date"
-          :data="patientsPerCities"
           :city-data-map="cityDataMap"
           :data-weekly="patientsPerCitiesWeekly"
-          :legends="patientsPerCitiesLegends"
           :title-date="patientsPerCitiesDate"
           :remarks="[
             '公表されている居住地が、54市町村以外(尾張地方、三河地方など)や県外は対象外',
@@ -127,7 +125,7 @@ import Data from '@/data/data.json'
 import CityData from '@/data/city_data.json'
 
 import formatConfirmedCases from '@/utils/formatConfirmedCases'
-import formatPatientsPerCities, {
+import {
   formatPatientsPerCitiesWeekly,
   makeCityDataMap
 } from '@/utils/formatPatientsPerCities'
@@ -174,12 +172,6 @@ export default {
 
     // 市区町村別の感染者数
     const cityDataMap = makeCityDataMap(CityData.data)
-    const patientsPerCitiesLegends = []
-    const patientsPerCities = formatPatientsPerCities(
-      CityData.data,
-      Data.patients.data,
-      patientsPerCitiesLegends
-    )
     const patientsPerCitiesDate = Data.patients_summary.data.slice(-1)[0][
       '日付'
     ]
@@ -192,9 +184,7 @@ export default {
       dataWeekly,
       confirmedCases,
       cityDataMap,
-      patientsPerCities,
       patientsPerCitiesWeekly,
-      patientsPerCitiesLegends,
       patientsPerCitiesDate,
       headerItem: {
         icon: 'mdi-chart-timeline-variant',
